requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
xero-python==9.3.0
requests==2.32.3
cryptography==43.0.3
python-dotenv==1.2.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
PyJWT>=2.8.0
itsdangerous==2.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
mysql-connector-python==9.1.0
phpserialize==1.3
